# Models/Todo.cs
using System.ComponentModel.DataAnnotations;

namespace ToDoApi.Models;

public class Todo
{
    [Key]
    public Guid Id { get; set; }
    public string Title { get; set; } = string.Empty;
    public string Description { get; set; } = string.Empty;
    public bool Completed { get; set; }
}